Extend and Super Keywords
We learned how to use the Extend and Super keywords in a constructor function to demonstrate inheritance between the functions. This allows us to remove duplicate code and share it from one super function with 'extend'.

Using 'super' allows us to use pieces of the super function inside a child function when defining new properties.

We also still have the flexibility to add additional properties on a single construction function outside of the super function.