Creating our first express app!
We're working on creating our fist node express app. We've installed express and started a server. We also learned about request/response and how to express's framework to do those actions by defining proper routes for the requests. We also made a route pattern and used the 'request' parameter to respond.