RESTful Routes continued
We implemented a commenting system that allows us to create and post comments to an index page. We also setup some redirects to get between the index, the comment form, and a details page for the comment.

Then we installed a new node module called UUID that creates a unique ID for each comments and setup a details button that creates an HTML page with a comment and all it's details including the unique ID.